refactor: decompose repos section into hook and subcomponents
